Refactorizacion del modelo de base de datos y del servicio remoto Corregidos errores menores
