+++ app/src/main/java/com/jesuslcorominas/posts/app/data/local/database/Mappers.kt
@@ -1,6 +1,6 @@
-package com.jesuslcorominas.posts.app.data.local.model
+package com.jesuslcorominas.posts.app.data.local.database
 
-import com.jesuslcorominas.posts.app.data.local.model.Post as DbPost
+import com.jesuslcorominas.posts.app.data.local.database.Post as DbPost
 import com.jesuslcorominas.posts.domain.Post as DomainPost
 
 fun DbPost.toDomainPost() = DomainPost(id, userId, title, body)
