+++ app/src/main/java/com/jesuslcorominas/posts/app/data/local/database/PostDatabase.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import com.jesuslcorominas.posts.app.data.local.model.Post
 
 @Database(entities = [Post::class], version = 1)
 abstract class PostDatabase : RoomDatabase() {
