+++ app/src/main/java/com/jesuslcorominas/posts/app/data/remote/service/Mappers.kt
@@ -1,6 +1,6 @@
-package com.jesuslcorominas.posts.app.data.remote.model
+package com.jesuslcorominas.posts.app.data.remote.service
 
-import com.jesuslcorominas.posts.app.data.remote.model.Post as RemotePost
+import com.jesuslcorominas.posts.app.data.remote.service.Post as RemotePost
 import com.jesuslcorominas.posts.domain.Post as DomainPost
 
 fun RemotePost.toDomainPost() = DomainPost(id, userId, title, body)
