+++ app/src/main/java/com/jesuslcorominas/posts/app/data/remote/service/RemoteApi.kt
@@ -1,6 +1,5 @@
 package com.jesuslcorominas.posts.app.data.remote.service
 
-import com.jesuslcorominas.posts.app.data.remote.model.Post
 import retrofit2.Call
 import retrofit2.http.GET
